requirements.txt
Django==4.2.2
djangorestframework==3.14.0
Markdown==3.4.3
django-filter==23.2
django-cors-headers==4.0.0
drf-yasg==1.21.5
django-extensions==3.2.3
psycopg2==2.9.6
django-redis==5.2.0
djangorestframework-simplejwt==5.2.2
Pillow==9.5.0